For crates from the same workspace, use path deps
This commit makes all dependencies that coem from the same workspace
`path` dependencies using the syntax:

```toml
actix-wamp = { path = "../actix-wamp", version = 0.2.0 }
```

Note that versioning info for local crates stays the same, however
the addition of `path` info makes the workspace and crates within
reusable if someone wants to link to the them via `git` or `path`
dependency. Also, if it ever so happens that one of the local crates
is not published but linked with the specified version number, the
project will no longer build stating it cannot resolve the dependency.
With this fix, this will no longer be a problem which is exactly
the problem right now.